+++ lib/beamer/locations/tasks_location.dart
@@ -1,47 +1,37 @@
 import 'package:beamer/beamer.dart';
 import 'package:flutter/material.dart';
-import 'package:lotti/pages/create/fill_survey_page.dart';
 import 'package:lotti/pages/create/record_audio_page.dart';
 import 'package:lotti/pages/journal/entry_details_page.dart';
 import 'package:lotti/pages/journal/infinite_journal_page.dart';
 import 'package:lotti/utils/uuid.dart';
 
-class JournalLocation extends BeamLocation<BeamState> {
-  JournalLocation(RouteInformation super.routeInformation);
+class TasksLocation extends BeamLocation<BeamState> {
+  TasksLocation(RouteInformation super.routeInformation);
 
   @override
   List<String> get pathPatterns => [
-        '/journal/:entryId',
-        '/journal/:entryId/record_audio/:linkedId',
-        '/journal/fill_survey/:surveyType',
+        '/tasks/:entryId',
+        '/tasks/:entryId/record_audio/:linkedId',
       ];
 
   @override
   List<BeamPage> buildPages(BuildContext context, BeamState state) {
     bool pathContains(String s) => state.uri.path.contains(s);
-    bool pathContainsKey(String s) => state.pathParameters.containsKey(s);
 
     final entryId = state.pathParameters['entryId'];
     final linkedId = state.pathParameters['linkedId'];
 
     return [
       const BeamPage(
-        key: ValueKey('journal'),
-        title: 'Journal',
+        key: ValueKey('tasks'),
+        title: 'Tasks',
         type: BeamPageType.noTransition,
-        child: InfiniteJournalPage(showTasks: false),
+        child: InfiniteJournalPage(showTasks: true),
       ),
       if (isUuid(entryId))
         BeamPage(
-          key: ValueKey('journal-$entryId'),
+          key: ValueKey('tasks-$entryId'),
           child: EntryDetailPage(itemId: entryId!),
-        ),
-      if (pathContains('fill_survey/') && pathContainsKey('surveyType'))
-        BeamPage(
-          key: ValueKey('fill_survey-${state.pathParameters['surveyType']}'),
-          child: FillSurveyWithTypePage(
-            surveyType: state.pathParameters['surveyType'],
-          ),
         ),
       if (pathContains('record_audio/'))
         BeamPage(