+++ lib/blocs/audio/recorder_cubit.dart
@@ -2,13 +2,13 @@ import 'dart:async';
 
 import 'package:bloc/bloc.dart';
 import 'package:intl/intl.dart';
-import 'package:lotti/beamer/beamer_delegates.dart';
 import 'package:lotti/blocs/audio/recorder_state.dart';
 import 'package:lotti/classes/audio_note.dart';
 import 'package:lotti/database/logging_db.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/logic/persistence_logic.dart';
 import 'package:lotti/services/asr_service.dart';
+import 'package:lotti/services/nav_service.dart';
 import 'package:lotti/utils/file_utils.dart';
 import 'package:lotti/utils/platform.dart';
 import 'package:record/record.dart';
@@ -105,7 +105,7 @@ class AudioRecorderCubit extends Cubit<AudioRecorderState> {
         );
         _linkedId = null;
 
-        journalBeamerDelegate.beamBack();
+        getIt<NavService>().beamBack();
         await getIt<AsrService>().transcribe(entry: entry!);
       }
     } catch (exception, stackTrace) {