+++ test/pages/journal/entry_detail_page_test.dart
@@ -18,6 +18,7 @@ import 'package:lotti/services/link_service.dart';
 import 'package:lotti/services/tags_service.dart';
 import 'package:lotti/services/time_service.dart';
 import 'package:lotti/themes/themes_service.dart';
+import 'package:lotti/utils/consts.dart';
 import 'package:lotti/widgets/journal/entry_tools.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:path_provider/path_provider.dart';
@@ -45,6 +46,11 @@ void main() {
         measurableChocolate,
       ]);
       mockPersistenceLogic = MockPersistenceLogic();
+
+      when(() => mockJournalDb.watchConfigFlag(enableTaskManagement))
+          .thenAnswer(
+        (_) => Stream<bool>.fromIterable([false]),
+      );
 
       final mockTagsService = mockTagsServiceWithTags([]);
       final mockTimeService = MockTimeService();